add This or That Recommendations and Settings UI , & update product results UI
This or That Recommendations UI is based on the UI design of the Categories UI for the Explore Categories product search option. THe maximum number of recommendations selectedable is 3. Selected items are highlighted in Green; non selected in Red. The product search results UI now contains relevant information about the product such as the image, name, and description. The skelton of the settings UI is created an included. All relavant views and modals utilizes kendo ui templates.