update the ThisorThat, and Home view with new icons
The Search field on the Home View has been refined to match the color of the header. The recommendation button's icon has been updated to a checkbox with a check. Plus, a badge is included to denote the number of recommendations selected. A check icon has been addedto the selected recommendations on the Modal View. "No Product Shown" messaging added to the Home View.