fix the event handlers, icons, styling, logic for this or that & home view
The event handlers for the Home View were removed because the ids weren't consitent. This or That View (Dual Search) has a No Recommendations Selected message when no selections are made. A check icon is available for confirming choices and the X dismisses the Recommendation modal. The This or That view's template issues which included problems dislaying the product items properly was fixed. The gift icons on the Home view associated to the list elements have event handlers for collecting the corresponding product name and ID.