update wishlist, login, and this or that view This or That View (Dual Search) THe product image was added to the this or that view.The header buttons' positions are reversed for the Recommendation modal view. Recommendation selection issue related to deselecting a recommendation is resolved. Two items are now visible on each page of the scroll view. This or That view is redesigned with no text descriptions or dividers. The ability to add items to the wishlist from this or that is available. A plus icon is added to the This and That button. The product retailer list view styling and display logic improved.
Login View
The functionality to reset the input field when the button group's buttons are tap is active.

Shares View
Shares view design implemented. Notification prompt present for sending wishlist request. Webservice not connected.

Overall
Touch responsiveness was improved for icons, buttons and other elements. A inner shadow style is applied when a button is pressed.